ci: optimize initial Cook tier CI scope
