+++ .github/actions/setup-cva6-env/action.yml
@@ -1,45 +1,39 @@
 # Copyright 2026 OpenHW Group
 # SPDX-License-Identifier: Apache-2.0
-#
-# Composite Action: Setup CVA6 Simulation Environment
-# Handles caching, optional tool installation, and environment setup for CVA6 CI jobs
+# CVA6 simulation tools and Cook dependencies.
 
-name: 'Setup CVA6 Environment'
-description: 'Cache/restore tools and prepare the CVA6 simulation environment'
+name: "Setup CVA6 Environment"
+description: "Cache/restore tools and prepare the CVA6 simulation environment"
 
 inputs:
   install-tools:
-    description: 'Run ci/setup.sh to install tools if cache miss'
+    description: "Run ci/setup.sh to install tools if cache miss"
     required: false
-    default: 'true'
+    default: "true"
   install-cook-dependencies:
-    description: 'Install the pinned Python environment required by cook.py'
+    description: "Install the pinned Python environment required by cook.py"
     required: false
-    default: 'false'
-  install-llvm18:
-    description: 'Install LLVM 18 tools required by Zcmt testlists'
-    required: false
-    default: 'false'
+    default: "false"
 
 outputs:
   toolchain-cache-hit:
-    description: 'Whether RISC-V toolchain cache was hit'
+    description: "Whether RISC-V toolchain cache was hit"
     value: ${{ steps.cache-toolchain.outputs.cache-hit }}
   verilator-cache-hit:
-    description: 'Whether Verilator cache was hit'
+    description: "Whether Verilator cache was hit"
     value: ${{ steps.cache-verilator.outputs.cache-hit }}
   spike-cache-hit:
-    description: 'Whether Spike cache was hit'
+    description: "Whether Spike cache was hit"
     value: ${{ steps.cache-spike.outputs.cache-hit }}
 
 runs:
-  using: 'composite'
+  using: "composite"
   steps:
     - name: Set up Python for cook.py
       if: inputs.install-cook-dependencies == 'true'
       uses: actions/setup-python@v6
       with:
-        python-version: '3.11'
+        python-version: "3.11"
         cache: pip
         cache-dependency-path: |
           flows/requirements.txt
@@ -53,19 +47,6 @@ runs:
           --constraint .github/requirements/cook-tier-ci-constraints.txt \
           --requirement flows/requirements.txt
 
-    - name: Install LLVM 18 for Zcmt testlists
-      if: inputs.install-llvm18 == 'true'
-      shell: bash
-      run: |
-        if ! command -v clang-18 >/dev/null || \
-           ! command -v llvm-objdump-18 >/dev/null || \
-           ! command -v llvm-nm-18 >/dev/null || \
-           ! command -v ld.lld-18 >/dev/null; then
-          sudo apt-get update
-          sudo apt-get install --yes clang-18 llvm-18 lld-18
-        fi
-
-    # Step 1: Get submodule hash for Spike cache key
     - name: Get core-v-verif submodule hash
       id: submodule-hash
       shell: bash
@@ -73,38 +54,27 @@ runs:
         cd verif/core-v-verif
         echo "hash=$(git rev-parse HEAD)" >> $GITHUB_OUTPUT
 
-    # Step 2: Cache RISC-V Toolchain
     - name: Cache RISC-V Toolchain
       id: cache-toolchain
       uses: actions/cache@v6
       with:
         path: tools/riscv-toolchain/
         key: ${{ runner.os }}-toolchain-${{ hashFiles('ci/install-toolchain.sh') }}
-        # restore-keys: |
-        #   ${{ runner.os }}-toolchain-
 
-    # Step 3: Cache Verilator
     - name: Cache Verilator
       id: cache-verilator
       uses: actions/cache@v6
       with:
         path: tools/verilator/
         key: ${{ runner.os }}-verilator-${{ hashFiles('verif/regress/install-verilator.sh') }}
-        # restore-keys: |
-        #   ${{ runner.os }}-verilator-
 
-    # Step 4: Cache Spike
     - name: Cache Spike
       id: cache-spike
       uses: actions/cache@v6
       with:
         path: tools/spike/
         key: ${{ runner.os }}-spike-${{ hashFiles('verif/regress/install-spike.sh') }}-${{ steps.submodule-hash.outputs.hash }}
-        # restore-keys: |
-        #   ${{ runner.os }}-spike-${{ hashFiles('verif/regress/install-spike.sh') }}-
-        #   ${{ runner.os }}-spike-
 
-    # Step 5: Install tools if cache miss and requested
     - name: Setup tools (if cache miss)
       if: |
         inputs.install-tools == 'true' && (
@@ -117,7 +87,6 @@ runs:
         set -x
         ci/setup.sh
 
-    # Step 6: Setup environment variables
     - name: Setup environment variables
       shell: bash
       run: |
