+++ .github/workflows/openhw-cva6-ci-tier2.yml
@@ -1,14 +1,8 @@
 # Copyright 2026 OpenHW Group
 # Licensed under the Apache License, Version 2.0, see LICENSE for details.
 # SPDX-License-Identifier: Apache-2.0
-# ================================================================================
-# OpenHW CVA6 CI - Tier 2: master_candidate cook.py regression
-# ================================================================================
-#
-# Broader public regression for the initial master_candidate AXI targets. Tier 2
-# expands the Tier 1 base tests with the supported Zcmt and PMP testlists while
-# preserving truthful job failures and downloadable evidence.
-# ================================================================================
+# Cook and Verilator tandem regression for master_candidate AXI targets.
+# Zcmt is deferred until the project LLVM version is agreed.
 
 name: openhw-cva6-ci-tier2
 
@@ -46,8 +40,8 @@ jobs:
           install-tools: "true"
           install-cook-dependencies: "true"
 
-      - name: Validate master_candidate Tier CI adapter
-        run: python3 .github/tests/test_master_candidate_cook_tier.py -v
+      - name: Test Verilator TestHarness result handling
+        run: python3 .github/tests/test_verilator_testharness_safety.py -v
 
   execute-rv32-tier2:
     name: RV32 Tier2 ${{ matrix.config }} / ${{ matrix.testcase }}
@@ -60,63 +54,21 @@ jobs:
         include:
           - config: cv32a60x_axi
             testcase: base-rv32-p
-            mode: cook-testlist
-            simulator: veri-testharness,spike
             install_script: install-riscv-tests
             testlist: verif/tests/base_rv32_p.yaml
-            toolchain: github_actions_gcc
             compiler_march: rv32imc_zicsr_zba_zbb_zbs_zbc
-            compiler_march_reason: public-gcc-13.2-does-not-accept-zcmt
-            expected_enabled_tests: 5
-            install_llvm18: false
-
-          - config: cv32a60x_axi
-            testcase: base-zcmt
-            mode: cook-testlist
-            simulator: veri-testharness,spike
-            install_script: ""
-            testlist: verif/tests/base_zcmt.yaml
-            toolchain: github_actions_llvm18
-            compiler_march: rv32imc_zicsr_zba_zbb_zbs_zbc_zcmt
-            compiler_march_reason: llvm-18-provides-zcmt-assembly-support
-            expected_enabled_tests: 4
-            install_llvm18: true
 
           - config: cv32a65x_axi
             testcase: base-rv32-p
-            mode: cook-testlist
-            simulator: veri-testharness,spike
             install_script: install-riscv-tests
             testlist: verif/tests/base_rv32_p.yaml
-            toolchain: github_actions_gcc
             compiler_march: rv32imc_zicsr_zba_zbb_zbs_zbc_zifencei
-            compiler_march_reason: public-gcc-13.2-does-not-accept-zcmt
-            expected_enabled_tests: 5
-            install_llvm18: false
-
-          - config: cv32a65x_axi
-            testcase: base-zcmt
-            mode: cook-testlist
-            simulator: veri-testharness,spike
-            install_script: ""
-            testlist: verif/tests/base_zcmt.yaml
-            toolchain: github_actions_llvm18
-            compiler_march: rv32imc_zicsr_zba_zbb_zbs_zbc_zcmt_zifencei
-            compiler_march_reason: llvm-18-provides-zcmt-assembly-support
-            expected_enabled_tests: 4
-            install_llvm18: true
 
           - config: cv32a65x_axi
             testcase: base-pmp
-            mode: cook-testlist
-            simulator: veri-testharness,spike
             install_script: ""
             testlist: verif/tests/base_pmp.yaml
-            toolchain: github_actions_gcc
             compiler_march: rv32imc_zicsr_zba_zbb_zbs_zbc_zifencei
-            compiler_march_reason: public-gcc-13.2-does-not-accept-zcmt
-            expected_enabled_tests: 5
-            install_llvm18: false
 
     steps:
       - uses: actions/checkout@v7
@@ -127,7 +79,6 @@ jobs:
         uses: ./.github/actions/setup-cva6-env
         with:
           install-cook-dependencies: "true"
-          install-llvm18: ${{ matrix.install_llvm18 }}
 
       - name: Install simulation prerequisites
         run: source ci/install-prereq.sh
@@ -135,16 +86,12 @@ jobs:
       - name: Run Tier 2 for ${{ matrix.config }} / ${{ matrix.testcase }}
         env:
           TIER_NAME: Tier 2
-          TIER_MODE: ${{ matrix.mode }}
+          TIER_MODE: cook-testlist
           TIER_CONFIG: ${{ matrix.config }}
           TIER_TESTCASE: ${{ matrix.testcase }}
-          TIER_SIMULATOR: ${{ matrix.simulator }}
           TIER_INSTALL_SCRIPT: ${{ matrix.install_script }}
           TIER_TESTLIST: ${{ matrix.testlist }}
-          TIER_TOOLCHAIN: ${{ matrix.toolchain }}
           TIER_COMPILER_MARCH: ${{ matrix.compiler_march }}
-          TIER_COMPILER_MARCH_REASON: ${{ matrix.compiler_march_reason }}
-          TIER_EXPECTED_ENABLED_TESTS: ${{ matrix.expected_enabled_tests }}
           TIER_EVENT_HEAD_SHA: ${{ github.sha }}
           TIER_EVENT_BASE_SHA: ${{ github.event.before || '' }}
         run: bash .github/scripts/run-tier-regression.sh
